requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.8.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.0.10
click==8.1.3
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==4.1.1
django-cors-headers==3.11.0
django-filter==22.1
djangorestframework==3.13.1
frozenlist==1.2.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
multidict==5.2.0
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.0.3
nest-asyncio==1.5.5
Pillow==9.0.0
protobuf==4.21.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rfc3986==1.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
tqdm==4.62.3
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.18.2
whitenoise==5.3.0
yarl==1.7.2
zope.interface==5.4.0
